src/main/java/za/co/wethinkcode/robotworlds/server/commands/BackCommand.java
package za.co.wethinkcode.robotworlds.server.commands;

import za.co.wethinkcode.robotworlds.server.world_bots.Robot;

public class BackCommand extends Command {

    public BackCommand(String argument) {
        super("back" , argument);
    }

    @Override
    public boolean execute(Robot target) {
        int nrSteps = Integer.parseInt(getArgument());
        if (target.updatePosition(-nrSteps)) {
            target.setStatus("Moved backward by " + nrSteps + " steps.");
        } else {
            target.setStatus("Sorry, I cannot go outside my safe zone.");
        }
        return true;
    }
}
